+++ makey/__init__.py
@@ -1,2 +1,2 @@
 # release version
-version = "1.3.0-dev1"
+version = "1.3.0-dev2"